Add Ticket Transaction History Filtering
This adds the ability to filter transactions in a ticket's history based
on the transaction type.  The set of filterable options is configurable
and defaults to 'Create', 'Correspond', and 'Comment'.  Queues can be
configured to be filtered by default.

The filtering is controlled by a new widget in the ticket transaction
history header that allows the filter to be turned on/off and for
individual (configured) transaction types to be toggled.