Provide default 'select all' query string
Some simple search terms are simplified down to an empty string, which
is usually construed as meaning "give me everything".  This change
provides a default in Results.html that should generally select
everything when no other criteria is provided.

This prevents displaying an "error parsing your search query" message
to the user when their seach term was factored away, or when they
provided an empty search term.