Add helper methods on Class for displaying articles
This change adds helper methods to RT::Class to assist in displaying an
article.

These methods provide a way to query the display configuration of the
class without having to check the attributes directly.  This covers the
display of the article name and summary, whether or not to EscapeHTML
on custom field values, which custom fields should be included, and
which custom fields should have their name displayed.

Also provides a convenience method to get an RT::CustomFields collection
object that is pre-filtered according to these display flags.