Update menus to respect config $UsernameFormat
Previously when building the menus RT always used the User->Name method
for the "Logged in as ..." menu.

While technically accurate, it isn't necessarily the most user-friendly,
particularly when External Authentication is in use and RT's "Name"
field may be set to something that's good for identification but that
the user may not use on a regular basis.

Using ->Format will respect User and System preferences.

A hidden component that's otherwise identical to the previous behavior
is added for those cases (hopefully just tests) where it's useful to
inspect the page contents for the ->Name of the current user.