Automatically load history for unread messages with on click setting
With the "click" option set for ticket history, a new message may not be
loaded initially when displaying a ticket. If the "Jump to Unread"
button is clicked, trigger the "click" of "Show ticket history" to load
history and jump to the corresponding transaction.

This also fixes another related issue that when the page has anchor to
some transaction id, browser didn't automatically jump after user clicks
"Show ticket history".