Convert RT to use GraphViz2
The GraphViz module has been delisted, making it unavailable for new
installs.

The GraphViz2 module does not present an identical API, so several
changes needed to be made.

As part of this update the graphs on /Ticket/Graphs/index.html are now
displayed as SVG with links instead of as PNG with a client-side
imagemap.

The PNG images are still available via the same paths as before this
change.