Fix bug in Logout menu option when using WebRemoteUserAuth
When WebRemoteUserAuth is enabled and WebFallbackToRTLogin is not
enabled the user would not see the Logout menu option.

Changed the logic to use the default RT logout when using default RT
login and only use the LogoutURL config option if the user is actually
authorized externally and the LogoutURL config option is not set to the
default RT logout.